+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.1.15] - 2021-12-05
+
 ### Fixed
 
 - Fixed twitch client token not refreshing